requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-bigquery==3.16.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==2.1.0
